test(commands): move `AnyUUID` mock in the `helpers` module
